Move tests inside directory
