+++ tests/backend/services/voting/FibonacciPointProvider.spec.ts
@@ -1,4 +1,4 @@
-import FibonacciPointProvider from "../../../backend/src/services/voting/FibonacciPointProvider";
+import FibonacciPointProvider from "../../../../backend/src/services/voting/FibonacciPointProvider";
 
 describe( "FibonacciPointProvider", () => {
 	let fibonacciPointProvider: FibonacciPointProvider;
